pyproject.toml
[tool.poetry]
name = "tomboone-com"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
azure-functions = "^1.21.3"
flask-login = "^0.6.3"
flask-sqlalchemy = "^3.1.1"
flask-wtf = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
